% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{add_component,i2dashboard,character-method}
\alias{add_component,i2dashboard,character-method}
\title{Method to add a component to a page of an i2dashboard.}
\usage{
\S4method{add_component}{i2dashboard,character}(dashboard, component, page = "default", copy = FALSE, ...)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{component}{The name of a function or a path to a file.}

\item{page}{The name of the page to add the component to.}

\item{copy}{Whether or not to copy images to \code{dashboard@datadir}.}

\item{...}{Additional parameters passed to the components render function.}
}
\description{
Components can be created by evaluating a function, or by including a text or image file.
The function tries to guess the intended usage by applying regular expressions to the argument \code{component} (see below for details).
}
\section{Adding a component by evaluating a function}{

If the argument \code{component} is a valid function name, the function will be called and its return value is used as component content.
}

\section{Adding plain text as a component}{

If the argument \code{component} ends with \code{.md} or \code{.txt}, the function will try to open a file and use its content as the components content.
}

\section{Adding images as a component}{

If the argument \code{component} end matches \code{\\.[png|jpg|jpeg|gif]}, the function will try to include an image as the components content.
}

