% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\docType{methods}
\name{add_component,i2dashboard,highchart-method}
\alias{add_component,i2dashboard,highchart-method}
\title{Method to add highcharts as components of a page of an i2dashboard object.}
\usage{
\S4method{add_component}{i2dashboard,highchart}(object, component,
  page = "default", title = NULL, ...)
}
\arguments{
\item{object}{The \linkS4class{i2dash::i2dashboard} object.}

\item{component}{The highcharts object.}

\item{page}{The name of the page to add the plot to.}

\item{title}{An optional component title.}
}
\description{
Method to add highcharts as components of a page of an i2dashboard object.
}
