% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_objects.R
\name{add_vis_object}
\alias{add_vis_object}
\title{General method to add an object as components of a page of an i2dashboard object.}
\usage{
add_vis_object(report, object, package, page = "default", title = NULL,
  ...)
}
\arguments{
\item{report}{The \linkS4class{i2dash::i2dashboard} object.}

\item{object}{The R visualization object to be addedd.}

\item{package}{The name of the R package that defines the class(object).}

\item{page}{The name of the page to add the object to.}

\item{title}{An optional component title.}
}
\description{
General method to add an object as components of a page of an i2dashboard object.
}
