% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and.R
\name{andUI}
\alias{andUI}
\title{AND module UI representation}
\usage{
andUI(id)
}
\arguments{
\item{id}{The ID of the modules namespace.}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
The AND module connects filtering and selection across multiple columns of a data.frame. Columns of class boolean, character or factor will be represented as textual ORs, numeric columns as numerical OR.
}
