% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPicker2.R
\name{categoricalPalettes}
\alias{categoricalPalettes}
\title{Function to generate categorical (qualitative) color palettes}
\usage{
categoricalPalettes(n)
}
\arguments{
\item{n}{Number of colors to generate}
}
\value{
A data.table with (named) color palettes of length n
}
\description{
Function to generate categorical (qualitative) color palettes
}
