% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarion.R
\docType{data}
\name{Clarion}
\alias{Clarion}
\title{Clarion R6-class definition}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Clarion
}
\arguments{
\item{header}{A named list. Defaults to NULL.}

\item{metadata}{Clarion metadata in form of a data.table.}

\item{data}{Data.table according to metadata.}

\item{validate}{Logical value to validate on initialization. Defaults to TRUE.}
}
\description{
Use this to create a clarion object.
This object is used by all top-level wilson modules.
}
\examples{
# initializing a new object
object <- Clarion$new(header, metadata, data, validate = TRUE)

# create a deep copy
object_copy <- object$clone(deep = TRUE)

}
\keyword{datasets}
