% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPicker.R
\name{colorPicker}
\alias{colorPicker}
\title{colorPicker module server logic}
\usage{
colorPicker(input, output, session)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}
}
\value{
The \code{input} object.
}
\description{
Provides server logic for the colorPicker module.
}
\section{To do}{

Implement transparency calculation in case of one or more single colors.
}

