% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPicker.R
\name{colorPickerUI}
\alias{colorPickerUI}
\title{colorPicker module UI representation}
\usage{
colorPickerUI(id, label = "Color scheme", choices = c("Blues", "Greens",
  "Greys", "Oranges", "Purples", "Reds"), selected.choice = NULL,
  show.reverse = TRUE, show.transparency = TRUE, single.colors = FALSE)
}
\arguments{
\item{id}{The ID of the modules namespace.}

\item{label}{Either a character vector of length one with the label for the color scheme dropdown, or a character vector containing labels of the single colors.}

\item{choices}{A character vector with choices for the color scheme dropdown. See \code{\link[shiny]{selectInput}}.}

\item{selected.choice}{The initially selected value(s) of the dropdown. If NULL (default), the first value of schemes will be taken.}

\item{show.reverse}{Logical, whether or not to show the reverse switch.}

\item{show.transparency}{Logical, whether or not to show the transparency slider.}

\item{single.colors}{Logical, whether or not to make a single color chooser. (Only if length(label) == 1 needed)}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
The functions creates HTML tag definitions of its representation based on the parameters supplied.
Currently, two UI can be created for the user to choose either (a) colors from a given color scheme, or (b) choose one or more single colors.
}
\section{Single color mode}{

If one or more single colors can be chosen, the UI element names are prefix by \emph{color} followed by \code{make.names} ouput of \code{label}.
}

\section{To do}{

Replace ordinary textInput for single colors by a real colorPicker, e.g. from https://github.com/daattali/colourpicker
}

