% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnSelector.R
\name{columnSelector}
\alias{columnSelector}
\title{columnSelector module server logic}
\usage{
columnSelector(input, output, session, type.columns, type = NULL,
  columnTypeLabel = "Type of Column", labelLabel = "Label",
  multiple = TRUE, none = FALSE, sep = ", ", suffix = NULL)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{type.columns}{data.table: (Supports reactive)
key = columnnames (id)
level = datalevel/ type of column
label = optional, used instead of id
sub_label = optional, added to id/ label}

\item{type}{The type (contrast/group/sample of the type dropdown menu, selected in step 1 (upper dropdown). Defaults to unique(type.columns[,2]) (Supports reactive)}

\item{columnTypeLabel}{Changes the label of the first UI element}

\item{labelLabel}{Change label above label text input.}

\item{multiple}{Boolean value whether multiple values can be selected in second selector. (Default = TRUE)}

\item{none}{If TRUE adds "None to secondSelector and select is. (Default = FALSE)}

\item{sep}{Used to seperate labels on multi value selection.}

\item{suffix}{Added to label only on multiple = FALSE (supports reactive). Also uses sep as seperator.}
}
\value{
Returns the input. As named list: names("type", "selectedColumns", "label")
}
\description{
columnSelector module server logic
}
