% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnSelector.R
\name{columnSelectorUI}
\alias{columnSelectorUI}
\title{columnSelector module UI representation}
\usage{
columnSelectorUI(id, label = FALSE, title = NULL)
}
\arguments{
\item{id}{The ID of the modules namespace.}

\item{label}{Boolean value; if true include a text input field with the desired axis label (this should be preset with the headline of the column)}

\item{title}{String which is displayed as module title. (Default = NULL)}
}
\value{
A list from \code{\link[shiny]{tag}} with the UI elements.
}
\description{
columnSelector module UI representation
}
