% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPicker.R
\name{divergingPalettes}
\alias{divergingPalettes}
\title{Function to generate diverging (two-sided) color palettes (e.g. for log2fc, zscore)}
\usage{
divergingPalettes(n)
}
\arguments{
\item{n}{Number of colors to generate}
}
\value{
A data.table with (named) color palettes of length n
}
\description{
Function to generate diverging (two-sided) color palettes (e.g. for log2fc, zscore)
}
