% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{.add_component}
\alias{.add_component}
\title{Helper function to add components to the dashboard}
\usage{
.add_component(dashboard, page, component)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{page}{The name of the page to add the component to.}

\item{component}{A string or list.}
}
\value{
The dashboard with added component.
}
\description{
Helper function to add components to the dashboard
}
