% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{.render_page}
\alias{.render_page}
\title{Method for rendering a page with a given layout and components}
\usage{
.render_page(
  title,
  components,
  layout = c("default", "storyboard", "focal_left", "2x2_grid"),
  menu = NULL,
  sidebar = NULL
)
}
\arguments{
\item{title}{The page title.}

\item{components}{A list of page components.}

\item{layout}{The pages overall layout.}

\item{menu}{The menu under which the page will be filed.}

\item{sidebar}{Character string with sidebar content.}
}
\value{
A markdown string with the final page.
}
\description{
Method for rendering a page with a given layout and components
}
