% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{download}
\alias{download}
\title{Function used for downloading.
Creates a zip container containing plot in png, pdf and user input in json format.
Use inside \code{\link[shiny]{downloadHandler}} content function.}
\usage{
download(file, filename, plot, width, height, ppi = 72, save_plot = TRUE,
  ui = NULL)
}
\arguments{
\item{file}{See \code{\link[shiny]{downloadHandler}} content parameter.}

\item{filename}{See \code{\link[shiny]{downloadHandler}}.}

\item{plot}{Plot to download.}

\item{width}{in centimeter.}

\item{height}{in centimeter.}

\item{ppi}{pixel per inch. Defaults to 72.}

\item{save_plot}{Logical if plot object should be saved as .RData.}

\item{ui}{List of user inputs. Will be converted to Javascript Object Notation. See \code{\link[RJSONIO]{toJSON}}}
}
\value{
See \code{\link[utils]{zip}}.
}
\description{
Function used for downloading.
Creates a zip container containing plot in png, pdf and user input in json format.
Use inside \code{\link[shiny]{downloadHandler}} content function.
}
