% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{embed_var}
\alias{embed_var}
\title{Method to download embed files into an Rmd-file}
\usage{
embed_var(x, ...)
}
\arguments{
\item{x}{Data, which will be written to the embedded file.}

\item{...}{Additional parameters.}
}
\description{
Method to download embed files into an Rmd-file
}
