% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{equalize}
\alias{equalize}
\title{Method to get equalized min/max values from vector}
\usage{
equalize(values)
}
\arguments{
\item{values}{Numeric vector or table}
}
\value{
Vector with c(min, max).
}
\description{
Method to get equalized min/max values from vector
}
