% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelector.R
\name{featureSelector}
\alias{featureSelector}
\title{featureSelector module server logic}
\usage{
featureSelector(input, output, session, clarion, multiple = TRUE,
  contains = FALSE, ranged = TRUE, step = 100, truncate = 30,
  selection.default = "all")
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}

\item{clarion}{A clarion object. See \code{\link[wilson]{Clarion}}. (Supports reactive)}

\item{multiple}{Whether or not textual ORs should allow multiple selections. (Fills vector sequentially if needed)(Supports reactive)}

\item{contains}{Whether or not textual ORs are initialized as textInput checking entries for given string. (Fills vector sequentially if needed)(Supports reactive)}

\item{ranged}{Whether or not numeric ORs are ranged. (Fills vector sequentially if needed)(Supports reactive)}

\item{step}{Set numeric ORs number of slider steps. (Fills vector sequentially if needed)(Supports reactive)}

\item{truncate}{Truncate datatable entries at x characters (Default = 30).}

\item{selection.default}{Decide whether everything or nothing is selected on default (no filters applied). Either "all" or "none" (Default = "all").}
}
\value{
Reactive containing names list: Selected data as reactive containing clarion object (object). Used filter to select data (filter).
}
\description{
featureSelector module server logic
}
\details{
Keep in mind that the order of features is the order in which delimiter, multiple, contains, ranged and step are evaluated.
}
