% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelector.R
\name{featureSelectorGuide}
\alias{featureSelectorGuide}
\title{featureSelector module guide}
\usage{
featureSelectorGuide(session, grouping = FALSE)
}
\arguments{
\item{session}{The shiny session}
}
\value{
A shiny reactive that contains the texts for the guide steps.
}
\description{
featureSelector module guide
}
