% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelector.R
\name{featureSelectorUI}
\alias{featureSelectorUI}
\title{featureSelector module UI representation}
\usage{
featureSelectorUI(id)
}
\arguments{
\item{id}{The ID of the modules namespace}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}
}
\description{
featureSelector module UI representation
}
