% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneView.R
\name{geneView}
\alias{geneView}
\title{geneView's module server logic}
\usage{
geneView(input, output, session, data, metadata, level = NULL,
  plot.method = "static", custom.label = NULL, label.sep = ", ",
  width = "auto", height = "auto", ppi = 72, scale = 1)
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}

\item{data}{data.table:
column1   : ids
column2   : symbol (data used for selection)
column3-n : data}

\item{metadata}{data.table:
column1: ids
column2: factor1 (conditions)
column3: level (condition type)}

\item{level}{Vector containing data levels to select from (default: unique(metadata[["level"]])).}

\item{plot.method}{Choose which method is used for plotting. Either "static" or "interactive" (Default = "static").}

\item{custom.label}{Data.table used for creating custom labels (supports reactive).}

\item{label.sep}{Seperator used for label merging (Default = ", ").}

\item{width}{Width of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{height}{Height of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{ppi}{Pixel per inch. Defaults to 72 and supports reactive.}

\item{scale}{Scale plot size. Defaults to 1, supports reactive.}
}
\value{
Reactive containing data.table used for plotting.
}
\description{
Provides server logic for the geneView module.
}
\details{
Width/ height/ ppi less or equal to default will use default value.

Ppi less or equal to zero will use default.
}
