% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneView.R
\name{geneViewGuide}
\alias{geneViewGuide}
\title{geneView module guide}
\usage{
geneViewGuide(session, label = FALSE)
}
\arguments{
\item{session}{The shiny session}

\item{label}{Boolean to show custom label step.}
}
\value{
A shiny reactive that contains the texts for the Guide steps.
}
\description{
geneView module guide
}
