% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneView.R
\name{geneViewUI}
\alias{geneViewUI}
\title{geneView's module UI representation}
\usage{
geneViewUI(id, plot.columns = 3)
}
\arguments{
\item{id}{The ID of the modules namespace.}

\item{plot.columns}{Initial value of plot column slider. Integer value between 1 and 7 (Default = 3).}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
geneView's module UI representation
}
