% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_cor_heatmap.R
\name{global_cor_heatmap}
\alias{global_cor_heatmap}
\title{global correlation heatmap module server logic}
\usage{
global_cor_heatmap(input, output, session, clarion,
  plot.method = "static", width = "auto", height = "auto",
  ppi = 72, scale = 1)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{clarion}{A clarion object. See \code{\link[wilson]{Clarion}}. (Supports reactive)}

\item{plot.method}{Choose which method is used for plotting. Either "static" or "interactive" (Default = "static").}

\item{width}{Width of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{height}{Height of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{ppi}{Pixel per inch. Defaults to 72 and supports reactive.}

\item{scale}{Scale plot size. Defaults to 1, supports reactive.}
}
\value{
Reactive containing data used for plotting.
}
\description{
global correlation heatmap module server logic
}
