% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_cor_heatmap.R
\name{global_cor_heatmap_guide}
\alias{global_cor_heatmap_guide}
\title{global correlation heatmap module guide}
\usage{
global_cor_heatmap_guide(session)
}
\arguments{
\item{session}{The shiny session}
}
\value{
A shiny reactive that contains the texts for the Guide steps.
}
\description{
global correlation heatmap module guide
}
