% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_cor_heatmap.R
\name{global_cor_heatmapUI}
\alias{global_cor_heatmapUI}
\title{global correlation heatmap module UI representation}
\usage{
global_cor_heatmapUI(id)
}
\arguments{
\item{id}{The ID of the modules namespace.}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}
}
\description{
global correlation heatmap module UI representation
}
