% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapUI}
\alias{heatmapUI}
\title{heatmap module UI representation}
\usage{
heatmapUI(id, row.label = TRUE)
}
\arguments{
\item{id}{The ID of the modules namespace.}

\item{row.label}{Boolean Value set initial Value for rowlabel checkbox (Default = TRUE).}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
heatmap module UI representation
}
