% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2dash-pkg.R
\docType{package}
\name{i2dash-package}
\alias{i2dash}
\alias{i2dash-package}
\title{'i2dash': A package for programmatic creation of interactive, web-based dashboards}
\description{
'i2dash' relies on the widely used R packages 'flexdashboard', 'knitr' and 'rmarkdown'. 'i2dash' introduces a new class from R's S4 object system named \linkS4class{i2dashboard}, which by design provides the main functionality of the package. Besides global properties such as the dashboard title, author and theme, an instance of the \linkS4class{i2dashboard} class also stores individual dashboard pages and the navigation menu, as well as all components that make up the content of individual pages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://loosolab.github.io/i2dash/}
  \item Report bugs at \url{https://github.com/loosolab/i2dash/issues}
}

}
\author{
\strong{Maintainer}: Arsenij Ustjanzew \email{arsenij.ustjanzew@gmail.com} (\href{https://orcid.org/0000-0002-1014-4521}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jens Preussner \email{jens.preussner@mpi-bn.mpg.de} (\href{https://orcid.org/0000-0003-1927-3458}{ORCID}) [copyright holder]
  \item Mario Looso \email{mario.looso@mpi-bn.mpg.de} (\href{https://orcid.org/0000-0003-1495-9530}{ORCID}) [copyright holder]
}

}
\keyword{internal}
