% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2dash-pkg.R
\docType{package}
\name{i2dash}
\alias{i2dash}
\title{'i2dash': A package for programmatic creation of interactive, web-based dashboards}
\description{
'i2dash' relies on the widely used R packages 'flexdashboard', 'knitr' and 'rmarkdown'. 'i2dash' introduces a new class from R's S4 object system named \linkS4class{i2dashboard}, which by design provides the main functionality of the package. Besides global properties such as the dashboard title, author and theme, an instance of the \linkS4class{i2dashboard} class also stores individual dashboard pages and the navigation menu, as well as all components that make up the content of individual pages.
}
\section{Citation}{


When using the package in your work, please cite:
i2dash: Creation of Flexible, Interactive and Web-based Dashboards for Visualization of Omics-pipeline Results
Arsenij Ustjanzew, Jens Preussner, Mette Bentsen, Carsten Kuenne, Mario Looso
bioRxiv 2020.07.06.189563; doi: https://doi.org/10.1101/2020.07.06.189563.
}

\author{
Arsenij Ustjanzew \email{arsenij.ustjanzew@gmail.com}

Jens Preussner \email{Jens.Preussner@mpi-bn.mpg.de}

Mario Looso \email{Mario.Looso@mpi-bn.mpg.de}
}
