% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\docType{methods}
\name{i2dashboard-class}
\alias{i2dashboard-class}
\title{Get/set the interactivity of the i2dashboard object.}
\usage{
\S4method{interactivity}{i2dashboard}(object)

\S4method{interactivity}{i2dashboard}(object) <- value
}
\arguments{
\item{value}{The value of the desired property.}

\item{.Object}{A \linkS4class{i2dash::i2dashboard} report.}
}
\description{
Get/set the interactivity of the i2dashboard object.
}
