% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R, R/components.R, R/pages.R
\docType{methods}
\name{assemble,i2dashboard-method}
\alias{assemble,i2dashboard-method}
\alias{add_component,i2dashboard,character-method}
\alias{add_page,i2dashboard-method}
\alias{remove_page,i2dashboard-method}
\title{Method to assemble a dashboard to a Rmd file.}
\usage{
\S4method{assemble}{i2dashboard}(object, pages = names(object@pages),
  file = object@file, render = FALSE, ...)

\S4method{add_component}{i2dashboard,character}(object, component,
  page = "default", ...)

\S4method{add_page}{i2dashboard}(object, page, title, layout = "default",
  menu = NULL, sidebar = NULL, ...)

\S4method{remove_page}{i2dashboard}(object, page)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{pages}{A string or vector with the names of pages, which should be assemble to a report.}

\item{file}{The output filename (recommend that the suffix should be '.Rmd'). This file will be saved in the working directory.}

\item{render}{A logical indicating whether the assembled report should immediately be rendered with \code{rmarkdown::render}.}

\item{...}{Additional arguments passed on to \code{rmarkdown::render}.}

\item{component}{The name of the component.}

\item{page}{The name of the page to add the component to.}

\item{title}{The title of the page to be added.}

\item{layout}{The page layout (see below).}

\item{menu}{The name of the menu, under which the page should appear.}

\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{...}{Additional parameters passed to the components render function.}

\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{page}{The name of the page to be added.}

\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{page}{The name of the page to be removed.}
}
\description{
Method to assemble a dashboard to a Rmd file.

Method to add a component to a page of an i2dashboard object

Method to add a page to an i2dashboard object

Method to remove a page to an i2dashboard object
}
