% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/pages.R
\name{add_page}
\alias{add_page}
\alias{remove_page}
\alias{add_page,i2dashboard-method}
\alias{remove_page,i2dashboard-method}
\title{Methods to add and remove pages of an i2dashboard.}
\usage{
add_page(dashboard, ...)

remove_page(dashboard, page)

\S4method{add_page}{i2dashboard}(
  dashboard,
  page,
  title,
  layout = "default",
  menu = NULL,
  sidebar = NULL,
  ...
)

\S4method{remove_page}{i2dashboard}(dashboard, page)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dashboard}.}

\item{...}{Additional arguments.}

\item{page}{The name of the page to be added or removed.}

\item{title}{The title of the page to be added.}

\item{layout}{The page layout (see below).}

\item{menu}{The name of the menu, under which the page should appear.}

\item{sidebar}{A Markdown string. Preferably, use the function add_to_sidebar.}
}
\value{
The \linkS4class{i2dashboard} object.
}
\description{
\code{add_page} creates a page and adds it to the \linkS4class{i2dashboard} object.
\code{remove_page} removes a page from the \linkS4class{i2dashboard} object.
}
