% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{label module server logic}
\usage{
label(
  input,
  output,
  session,
  data,
  label = "Select label columns",
  multiple = TRUE,
  sep = ", ",
  unique = TRUE,
  unique_sep = "_",
  disable = NULL
)
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}

\item{data}{Data.table used for label creation. Column names will be used for selection. (supports reactive)}

\item{label}{Set label of selectizeInput.}

\item{multiple}{Allow multiple selection which will be merged with sep (default = TRUE).}

\item{sep}{Separator used to collapse selection (default = ", ").}

\item{unique}{Make labels unique. Defaults to TRUE. See \code{\link[base]{make.unique}}.}

\item{unique_sep}{Separator used for unique (default = "_"). Should differ from sep.}

\item{disable}{Reactive containing boolean. To disable/ enable module.}
}
\value{
Reactive containing list(label = vector of strings or NULL on empty selection, selected = user input).
}
\description{
label module server logic
}
