% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{labelUI}
\alias{labelUI}
\title{label module UI representation}
\usage{
labelUI(id)
}
\arguments{
\item{id}{The ID of the modules namespace}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}
}
\description{
label module UI representation
}
