% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit.R
\name{limit}
\alias{limit}
\title{limit module server logic}
\usage{
limit(input, output, session, lower = NULL, upper = NULL)
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}

\item{lower}{Set lower limit (supports reactive).}

\item{upper}{Set upper limit (supports reactive).}
}
\value{
Reactive containing: list(lower, upper).
}
\description{
limit module server logic
}
