% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit.R
\name{limitUI}
\alias{limitUI}
\title{limit module UI representation}
\usage{
limitUI(id, label = "Limit")
}
\arguments{
\item{id}{The ID of the modules namespace}

\item{label}{Set the modules label.}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}
}
\description{
limit module UI representation
}
