% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{marker}
\alias{marker}
\title{marker module server logic}
\usage{
marker(input, output, session, highlight.labels)
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}

\item{highlight.labels}{Data.table from which labels are provided (Supports reactive).}
}
\value{
A reactive which contains a named list (highlight, color, labelColumn, label).
}
\description{
marker module server logic
}
