% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{markerUI}
\alias{markerUI}
\title{marker module UI representation}
\usage{
markerUI(id, label = "Highlight/ Label Selected Features")
}
\arguments{
\item{id}{The ID of the modules namespace}

\item{label}{Set label of first element.}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}
}
\description{
marker module UI representation
}
