% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orNumeric.R
\name{orNumericUI}
\alias{orNumericUI}
\title{orNumeric module UI representation}
\usage{
orNumericUI(id)
}
\arguments{
\item{id}{The ID of the modules namespace.}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
This module allows to select value/range inputs from a \code{\link[shiny]{sliderInput}} element.
The functions creates HTML tag definitions of its representation based on the parameters supplied.
}
