% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orTextual.R
\name{orTextual}
\alias{orTextual}
\title{orTextual module server logic}
\usage{
orTextual(input, output, session, choices, selected = NULL,
  label = "Column", delimiter = NULL, multiple = TRUE,
  contains = FALSE, reset = NULL)
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}

\item{choices}{A list or a character vector with the possible choices offered in the UI. See \code{\link[shiny]{selectInput}}.}

\item{selected}{The initially selected value. See \code{\link[shiny]{selectInput}}.}

\item{label}{A character vector of length one with the label for the \code{\link[shiny]{selectInput}}.}

\item{delimiter}{A single character indicating if and how items are delimited (default: \code{NULL} indicates no delimitation). Only if contains = FALSE.}

\item{multiple}{Whether or not selection of multiple items is allowed.}

\item{contains}{Logical variable. If TRUE shows module as a textsearch input.}

\item{reset}{A reactive which will trigger a module reset on change.}
}
\value{
Returns a reactive containing a named list with the label, the selected choices as a character vector (text) and a boolean vector of length \code{length(choices)} (bool), indicating whether a item has been chosen. If no item has been chosen, the return is \code{TRUE} for items.
}
\description{
Provides server logic for the orTextual module.
}
