% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orTextual.R
\name{orTextualUI}
\alias{orTextualUI}
\title{orTextual module UI representation}
\usage{
orTextualUI(id)
}
\arguments{
\item{id}{The ID of the modules namespace.}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
This module allows to select (multiple) inputs from a \code{\link[shiny]{selectInput}} element.
The functions creates HTML tag definitions of its representation based on the parameters supplied.
}
