% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser}
\alias{parser}
\title{Method to parse input file.}
\usage{
parser(file, dec = ".")
}
\arguments{
\item{file}{Path to file that needs parsing.}

\item{dec}{The decimal separator. See \code{\link[data.table]{fread}}.}
}
\value{
Clarion object. See \code{\link[wilson]{Clarion}}
}
\description{
Method to parse input file.
}
