% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{pca module server logic}
\usage{
pca(
  input,
  output,
  session,
  clarion,
  width = 28,
  height = 28,
  ppi = 72,
  scale = 1
)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{clarion}{A clarion object. See \code{\link[wilson]{Clarion}}. (Supports reactive)}

\item{width}{Width of the plot in cm. Defaults to 28 and supports reactive.}

\item{height}{Height of the plot in cm. Defaults to 28 and supports reactive.}

\item{ppi}{Pixel per inch. Defaults to 72 and supports reactive.}

\item{scale}{Scale plot size. Defaults to 1, supports reactive.}
}
\value{
A reactive containing list with dimensions.
}
\description{
pca module server logic
}
\details{
Width/ height/ ppi less or equal to zero will use default value.
}
