% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaGuide}
\alias{pcaGuide}
\title{pca module guide}
\usage{
pcaGuide(session)
}
\arguments{
\item{session}{The shiny session}
}
\value{
A shiny reactive that contains the texts for the Guide steps.
}
\description{
pca module guide
}
