% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaUI}
\alias{pcaUI}
\title{pca module UI representation}
\usage{
pcaUI(id, show.label = TRUE)
}
\arguments{
\item{id}{The ID of the modules namespace.}

\item{show.label}{Set initial value of show label checkbox (Default = TRUE).}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
pca module UI representation
}
