% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_questions.R
\name{release_questions}
\alias{release_questions}
\title{Defines additional questions asked before CRAN submission.
DO NOT EXPORT!}
\usage{
release_questions()
}
\description{
Defines additional questions asked before CRAN submission.
DO NOT EXPORT!
}
