% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_by_factors.R
\name{render_features_by_factors}
\alias{render_features_by_factors}
\title{Renders a features by factor violin plot}
\usage{
render_features_by_factors(object, plot_title, x, y)
}
\arguments{
\item{plot_title}{The title of the Component}

\item{x}{A list with the x-axis values. If it is a nested list, a dropdown-field will be provided in the interactive mode.}

\item{y}{A list with the y-axis values. If it is a nested list, a dropdown-field will be provided in the interactive mode.(Needs to be categorial. Horizontal violinplots are not possible.)}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a features by factor violin plot
}
