% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{render_image}
\alias{render_image}
\title{Method to embed an image file in a component}
\usage{
render_image(
  image,
  image_alt_text = NULL,
  title = NULL,
  raw = FALSE,
  width = "100\%",
  height = "auto",
  in_component = TRUE
)
}
\arguments{
\item{image}{The path to the image file.}

\item{image_alt_text}{The alt text of the image.}

\item{title}{The components title.}

\item{raw}{Whether or not to emit solely the markdown image code.}

\item{width}{Width defined with CSS in the HTML img-tag.}

\item{height}{Height defined with CSS in the HTML img-tag.}

\item{in_component}{Whether the image belongs to a component or the sidebar}
}
\value{
A character string containing the evaluated component
}
\description{
Method to embed an image file in a component
}
