% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{render_multiplot}
\alias{render_multiplot}
\title{Renders a Sequence saturation plot}
\usage{
render_multiplot(object, plot_title, x, y, color_by)
}
\arguments{
\item{plot_title}{The title of the Component}

\item{x}{A list with the x-axis values. If it is a nested list, a dropdown-field will be provided in the interactive mode.}

\item{y}{A list with the y-axis values. If it is a nested list, a dropdown-field will be provided in the interactive mode.}

\item{color_by}{A list with the color_by values. If it is a nested list, a dropdown-field will be provided in the interactive mode.}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a Sequence saturation plot
}
