% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{render_text}
\alias{render_text}
\title{Method to embed content from a text file in a component}
\usage{
render_text(file, title = NULL)
}
\arguments{
\item{file}{The file containing the text content.}

\item{title}{The components title.}
}
\value{
A character string containing the evaluated component
}
\description{
Method to embed content from a text file in a component
}
