% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textbox.R
\name{render_textbox}
\alias{render_textbox}
\title{Renders a textbox with arbitrary content}
\usage{
render_textbox(title, content)
}
\arguments{
\item{title}{The title of the textbox}

\item{content}{The content of the textbox}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Renders a textbox with arbitrary content
}
