% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{scatterPlot module server logic}
\usage{
scatterPlot(input, output, session, data, types, features = NULL,
  markerReac = NULL, plot.method = "static", width = "auto",
  height = "auto", ppi = 72, scale = 1)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{data}{data.table data visualized in plot (Supports reactive).}

\item{types}{data.table: (Supports reactive)
column1: colnames of data
column2: corresponding column type
column3: label (optional, used instead of id)
column4: sub_label (optional, added to id/ label)}

\item{features}{data.table of the features to mark (first column = id)}

\item{markerReac}{reactive containing inputs of marker module.}

\item{plot.method}{Choose to rather render a 'interactive' or 'static' plot. Defaults to 'static'.}

\item{width}{Width of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{height}{Height of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{ppi}{Pixel per inch. Defaults to 72 and supports reactive.}

\item{scale}{Scale plot size. Defaults to 1, supports reactive.}
}
\value{
Returns reactive containing data used for plot.
}
\description{
scatterPlot module server logic
}
\details{
Make sure to have the same columnnames in data and features.
}
