% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlotGuide}
\alias{scatterPlotGuide}
\title{scatterPlot module guide}
\usage{
scatterPlotGuide(session, marker = FALSE)
}
\arguments{
\item{session}{The shiny session}

\item{marker}{Logical if marker step should be enabled (Default = FALSE).}
}
\value{
A shiny reactive that contains the texts for the Guide steps.
}
\description{
scatterPlot module guide
}
