% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlotUI}
\alias{scatterPlotUI}
\title{scatterPlot module UI representation}
\usage{
scatterPlotUI(id)
}
\arguments{
\item{id}{The ID of the modules namespace.}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
scatterPlot module UI representation
}
