% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPicker2.R
\name{sequentialPalettes}
\alias{sequentialPalettes}
\title{Function to generate sequential (one-sided) color palettes (e.g. for expression, enrichment)}
\usage{
sequentialPalettes(n)
}
\arguments{
\item{n}{Number of colors to generate}
}
\value{
A data.table with (named) color palettes of length n
}
\description{
Function to generate sequential (one-sided) color palettes (e.g. for expression, enrichment)
}
