% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{tobias_parser}
\alias{tobias_parser}
\title{TOBIAS TFBS table to clarion parser}
\usage{
tobias_parser(input, output, filter_columns = NULL,
  filter_pattern = NULL, config = system.file("extdata",
  "tobias_config.json", package = "wilson"), omit_NA = FALSE,
  condition_names = NULL, condition_pattern = "_bound$",
  in_field_delimiter = ",", dec = ".", ...)
}
\arguments{
\item{input}{Path to input table}

\item{output}{Output path.}

\item{filter_columns}{Either a vector of columnnames or a file containing one columnname per row.}

\item{filter_pattern}{Keep columns matching the given pattern. Parameter filter_columns will be combined with the actual columnnames from the table before pattern is matched. In the case of no matches a warning will be issued and all columns will be used.}

\item{config}{Json file containing metadata information for all columns.}

\item{omit_NA}{Logical whether all rows containing NA should be removed.}

\item{condition_names}{Vector of condition names. Default = NULL. Used to classify columns not provided in config.}

\item{condition_pattern}{Used to identify condition names by matching und removing given pattern with \code{\link[base]{grep}}. Ignored when condition_names is set.}

\item{in_field_delimiter}{Delimiter for multi value fields. Default = ','.}

\item{dec}{Decimal separator. Used in file reading and writing.}

\item{...}{Used as header information.}
}
\description{
TOBIAS TFBS table to clarion parser
}
