% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transformation}
\alias{transformation}
\title{transformation module server logic}
\usage{
transformation(input, output, session, data, transpose = FALSE,
  pseudocount = 1, replaceInf = TRUE, replaceNA = TRUE)
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}

\item{data}{Numeric matrix on which transformation is performed (column-wise). (Supports reactive)}

\item{transpose}{Whether the matrix should be transposed to enable row-wise transformation.}

\item{pseudocount}{Numeric Variable to add a pseudocount to log-based transformations.}

\item{replaceInf}{Change Infinite to NA, applied after transformation.}

\item{replaceNA}{Change NA to 0, applied after transformation.}
}
\value{
Namedlist of two containing data and name of the used method.
        data: Reactive containing the transformed matrix. Infinite values are replaced by NA and NA values are replaced by 0.
        method: Reactive containing String.
        transpose: Reactive containing String.
}
\description{
The module provides several transformations on a numeric data matrix for the user.
}
