% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transformationUI}
\alias{transformationUI}
\title{transformation module UI representation}
\usage{
transformationUI(id, label = "Transformation", selected = "raw",
  choices = list(None = "raw", log2 = "log2", `-log2` = "-log2", log10 =
  "log10", `-log10` = "-log10", `Z score` = "zscore", `regularized log` =
  "rlog"), transposeOptions = FALSE)
}
\arguments{
\item{id}{The ID of the modules namespace.}

\item{label}{A character vector of length one with the label for the \code{\link[shiny]{selectInput}}.}

\item{selected}{The initially selected value. See \code{\link[shiny]{selectInput}}.}

\item{choices}{Named list of available transformations. Possible transformations are list(`None` = "raw", `log2` = "log2", `-log2` = "-log2", `log10` = "log10", `-log10` = "-log10", `Z score` = "zscore", `regularized log` = "rlog") which is also the default.}

\item{transposeOptions}{Boolean value if transpose radioButtons are shown (Default = FALSE).}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
This function provides an input to select a transformation method.
}
