% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{add_component,i2dashboard,character-method}
\alias{add_component,i2dashboard,character-method}
\title{Method to add a component to a page of an i2dashboard.}
\usage{
\S4method{add_component}{i2dashboard,character}(dashboard, component, page = "default", copy = FALSE, ...)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{component}{A R object, function, or character.}

\item{page}{The name of the page to add the component to.}

\item{copy}{Whether or not to copy images to \code{dashboard@datadir}.}

\item{...}{Additional parameters passed to the components render function.}
}
\description{
Components can be created by evaluating a function, or by including an object, a text or image file.
}
\section{Adding content by evaluating a function}{

If the argument \code{component} is a function, the function will be called and its return value is used as content.
}

\section{Adding plain text as content}{

If the argument \code{component} is a character and ends with \code{.md} or \code{.txt}, the function will try to open a file and use its content.
}

\section{Adding images as content}{

If the argument \code{component} is a character and its end matches \code{\\.[png|jpg|jpeg|gif]}, the function will try to include an image as the content.
}

\section{Adding a R object as content}{

If the argument \code{component} is a supported R object (e.g. a htmlwidget), the function will include its representation as content.
}

