% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_objects.R
\name{add_vis_object}
\alias{add_vis_object}
\title{General method to add an object as component to a page of an \linkS4class{i2dashboard} object.}
\usage{
add_vis_object(dashboard, object, package, page = "default", title = NULL)
}
\arguments{
\item{dashboard}{The \linkS4class{i2dashboard}.}

\item{object}{The R visualization object to be added.}

\item{package}{The name of the R package that defines the class(object).}

\item{page}{The name of the page to add the object to.}

\item{title}{An optional component title.}
}
\value{
Returns the modified \linkS4class{i2dashboard} object, if the component was added, otherwise an unmodified \linkS4class{i2dashboard} object.
}
\description{
General method to add an object as component to a page of an \linkS4class{i2dashboard} object.
}
