% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/assemble.R
\name{assemble}
\alias{assemble}
\alias{assemble,i2dashboard-method}
\title{Generate an RMarkdown file from an i2dashboard object.}
\usage{
assemble(dashboard, ...)

\S4method{assemble}{i2dashboard}(
  dashboard,
  pages = names(dashboard@pages),
  file = dashboard@file,
  exclude = NULL,
  render = FALSE,
  ...
)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dashboard}.}

\item{...}{Additional arguments passed on to \code{rmarkdown::render}.}

\item{pages}{A string or vector with the names of pages, which should be assembled into the resulting Rmd file.}

\item{file}{The filename of the resulting Rmd file (recommend that the suffix should be '.Rmd').}

\item{exclude}{A string or vector with the names of pages, which should be excluded from dashboard assembly.}

\item{render}{A logical indicating whether the assembled dashboard should immediately be rendered with \code{rmarkdown::render} or run with \code{rmarkdown::run}.}
}
\value{
Invisibly returns the dashboard.
}
\description{
Generate an RMarkdown file from an i2dashboard object.
}
