% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPicker.R
\name{colorPicker}
\alias{colorPicker}
\title{colorPicker module server logic}
\usage{
colorPicker(input, output, session, num.colors = 256, distribution = "all",
  winsorize = NULL, selected = NULL)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{num.colors}{Define length of colorpalette vector (Default = 256).}

\item{distribution}{Decide which palettes are selectable. One or more of list("sequential", "diverging", "categorical"). Defaults to "all" (Supports reactive).}

\item{winsorize}{Numeric vector of two. Dynamically change lower and upper limit (supports reactive). Defaults to NULL.}

\item{selected}{Set the default selected palette.}
}
\value{
Reactive containing list(palette = c(colors), name = palette_name, transparency = Integer, reverse = Boolean, winsorize = NULL or a two-component vector containing lower and upper limits).
}
\description{
Provides server logic for the colorPicker2 module.
}
\details{
A custom colorpalette's return will be NULL if there is something wrong with it.

equalize will be returned as FALSE if not selected.
}
