% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPicker.R
\name{colorPickerUI}
\alias{colorPickerUI}
\title{colorPicker module UI representation}
\usage{
colorPickerUI(
  id,
  label = "Color scheme",
  custom = FALSE,
  multiple = FALSE,
  show.reverse = TRUE,
  show.scaleoptions = TRUE,
  show.transparency = TRUE
)
}
\arguments{
\item{id}{The ID of the modules namespace.}

\item{label}{Either a character vector of length one with the label for the color scheme dropdown, or a character vector containing labels of the single colors.}

\item{custom}{Boolean if TRUE custom colors can be selected (Default = FALSE).}

\item{multiple}{Boolean value, if set to TRUE custom colorpalettes can be made. Only if custom = TRUE (Default = FALSE).}

\item{show.reverse}{Logical, whether or not to show the reverse switch (Default = TRUE).}

\item{show.scaleoptions}{Logical, whether or not to show color scaling option winsorize (Default = TRUE).}

\item{show.transparency}{Logical, whether or not to show the transparency slider (Default = TRUE).}
}
\value{
A list with HTML tags from \code{\link[shiny]{tag}}.
}
\description{
The functions creates HTML tag definitions of its representation based on the parameters supplied.
Currently, two UI can be created for the user to choose either (a) colors from a given color scheme, or (b) choose one or more single colors.
}
