% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{embed_var}
\alias{embed_var}
\title{A method to embed tabular data into an HTML link for download.}
\usage{
embed_var(x, ...)
}
\arguments{
\item{x}{Data, which will be written to the embedded file.}

\item{...}{Additional parameters.}
}
\description{
A method to embed tabular data into an HTML link for download.
}
\examples{
embed_var(mtcars)
}
