% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_cor_heatmap.R
\name{global_cor_heatmap}
\alias{global_cor_heatmap}
\title{global correlation heatmap module server logic}
\usage{
global_cor_heatmap(input, output, session, data, types,
  plot.method = "static", width = "auto", height = "auto", ppi = 72,
  scale = 1)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{data}{data.table data visualized in plot (supports reactive).}

\item{types}{data.table: (supports reactive)
column1: colnames of data
column2: corresponding column type
column3 = label (optional, used instead of id)
column4 = sub_label (optional, added to id/ label)}

\item{plot.method}{Choose which method is used for plotting. Either "static" or "interactive" (Default = "static").}

\item{width}{Width of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{height}{Height of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{ppi}{Pixel per inch. Defaults to 72 and supports reactive.}

\item{scale}{Scale plot size. Defaults to 1, supports reactive.}
}
\value{
Reactive containing data used for plotting.
}
\description{
global correlation heatmap module server logic
}
