% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{heatmap module server logic}
\usage{
heatmap(input, output, session, data, types, plot.method = "static",
  custom.row.label = NULL, label.sep = ", ", width = "auto",
  height = "auto", ppi = 72)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{data}{data.table data visualized in plot (Supports reactive).}

\item{types}{data.table: (Supports reactive)
column1: colnames of data
column2: corresponding column typ
column3: label (optional, used instead of id)
column4: sub_label (optional, added to id/ label)}

\item{plot.method}{Choose which method is used for plotting. Either "static" or "interactive" (Default = "static").}

\item{custom.row.label}{Data.table used for creating custom labels (supports reactive).}

\item{label.sep}{Seperator used for label merging (Default = ", ").}

\item{width}{Width of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{height}{Height of the plot in cm. Defaults to minimal size for readable labels and supports reactive.}

\item{ppi}{Pixel per inch. Defaults to 72 and supports reactive.}
}
\value{
Reactive containing data used for plotting.
}
\description{
heatmap module server logic
}
