% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2dash.R
\name{heatmap_to_i2dash}
\alias{heatmap_to_i2dash}
\title{Prepare a heatmap to be rendered with the i2dash package.}
\usage{
heatmap_to_i2dash(object, countTable, group_by, title = NULL, ...)
}
\arguments{
\item{object}{A \code{\link[i2dash]{i2dashboard}} object.}

\item{countTable}{A matrix with features as rows and observations as columns. The rownames and columnnames should be provided and are used in building the heatmap.}

\item{group_by}{A vector with numerical values or a named list will be mapped to the y-axis. In case of a named list, a dropdown menu will be provided in the interactive mode. Note: The length of vectors x and y should be the same as well as the length of all vectors in case of a named list.}

\item{title}{(Optional) The title of the components chunk.}

\item{...}{Further parameters which are compatible with wilsons create_heatmap() method. See \code{\link{create_heatmap}}.}
}
\value{
A string containing markdown code for the rendered textbox
}
\description{
Prepare a heatmap to be rendered with the i2dash package.
}
