% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapGuide}
\alias{heatmapGuide}
\title{heatmap module guide}
\usage{
heatmapGuide(session)
}
\arguments{
\item{session}{The shiny session}
}
\value{
A shiny reactive that contains the texts for the Guide steps.
}
\description{
heatmap module guide
}
