% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2dashboard.R
\docType{package}
\name{i2dash}
\alias{i2dash}
\title{i2dash: A package for programmatic creation of interactive, web-based dashboards}
\description{
i2dash relies on the widely used R packages flexdashboard, knitr and rmarkdown. i2dash introduces a new class from R's S4 object system named \linkS4class{i2dashboard}, which by design provides the main functionality of the package. Besides global properties such as the dashboard title, author and theme, an instance of the \linkS4class{i2dashboard} class also stores individual dashboard pages and the navigation menu, as well as all components that make up the content of individual pages.
}
\section{Citation}{


When using the package in your work, please cite: tba.
}

