% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set.R
\name{i2dashboard-class}
\alias{i2dashboard-class}
\title{Get/set the interactivity of the i2dashboard.}
\usage{
\S4method{interactivity}{i2dashboard}(dashboard)

\S4method{interactivity}{i2dashboard}(dashboard) <- value
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{value}{The value of the desired property.}
}
\description{
Get/set the interactivity of the i2dashboard.
}
