% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_set.R
\name{interactivity}
\alias{interactivity}
\alias{interactivity<-}
\alias{title}
\alias{title<-}
\alias{author}
\alias{author<-}
\alias{theme}
\alias{theme<-}
\alias{datadir}
\alias{datadir<-}
\alias{share}
\alias{share<-}
\alias{embed_source<-}
\alias{source}
\alias{source<-}
\alias{interactivity,i2dashboard-method}
\alias{interactivity<-,i2dashboard-method}
\alias{title,i2dashboard-method}
\alias{title<-,i2dashboard-method}
\alias{author,i2dashboard-method}
\alias{author<-,i2dashboard-method}
\alias{theme,i2dashboard-method}
\alias{theme<-,i2dashboard-method}
\alias{datadir,i2dashboard-method}
\alias{datadir<-,i2dashboard-method}
\alias{share,i2dashboard-method}
\alias{share<-,i2dashboard-method}
\alias{source,i2dashboard-method}
\alias{source<-,i2dashboard-method}
\alias{embed_source<-,i2dashboard-method}
\title{Accessor methods for slots of an \linkS4class{i2dashboard} object.}
\usage{
interactivity(dashboard)

interactivity(dashboard) <- value

title(dashboard)

title(dashboard) <- value

author(dashboard)

author(dashboard) <- value

theme(dashboard)

theme(dashboard) <- value

datadir(dashboard)

datadir(dashboard) <- value

share(dashboard)

share(dashboard) <- value

embed_source(dashboard) <- value

source(dashboard)

source(dashboard) <- value

\S4method{interactivity}{i2dashboard}(dashboard)

\S4method{interactivity}{i2dashboard}(dashboard) <- value

\S4method{title}{i2dashboard}(dashboard)

\S4method{title}{i2dashboard}(dashboard) <- value

\S4method{author}{i2dashboard}(dashboard)

\S4method{author}{i2dashboard}(dashboard) <- value

\S4method{theme}{i2dashboard}(dashboard)

\S4method{theme}{i2dashboard}(dashboard) <- value

\S4method{datadir}{i2dashboard}(dashboard)

\S4method{datadir}{i2dashboard}(dashboard) <- value

\S4method{share}{i2dashboard}(dashboard)

\S4method{share}{i2dashboard}(dashboard) <- value

\S4method{source}{i2dashboard}(dashboard)

\S4method{source}{i2dashboard}(dashboard) <- value

\S4method{embed_source}{i2dashboard}(dashboard) <- value
}
\arguments{
\item{dashboard}{A \linkS4class{i2dashboard}.}

\item{value}{The value of the desired property. See \linkS4class{i2dashboard} for details.}
}
\description{
Getter and Setter methods can be used to directly manipulate properties (slots) of an \linkS4class{i2dashboard} object. See \linkS4class{i2dashboard} for details.
}
