% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R, R/colormap.R, R/pages.R,
%   R/sidebar.R
\docType{methods}
\name{assemble,i2dashboard-method}
\alias{assemble,i2dashboard-method}
\alias{add_colormap,i2dashboard-method}
\alias{add_page,i2dashboard-method}
\alias{remove_page,i2dashboard-method}
\alias{add_to_sidebar,i2dashboard-method}
\title{Method to assemble a dashboard to a Rmd file.}
\usage{
\S4method{assemble}{i2dashboard}(dashboard,
  pages = names(dashboard@pages), file = dashboard@file,
  exclude = NULL, render = FALSE, ...)

\S4method{add_colormap}{i2dashboard}(dashboard, map, name)

\S4method{add_page}{i2dashboard}(dashboard, page, title,
  layout = "default", menu = NULL, sidebar = NULL, ...)

\S4method{remove_page}{i2dashboard}(dashboard, page)

\S4method{add_to_sidebar}{i2dashboard}(dashboard, content,
  page = "default", global = FALSE, copy = FALSE, ...)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{pages}{A string or vector with the names of pages, which should be assembled to a report.}

\item{file}{The output filename (recommend that the suffix should be '.Rmd'). This file will be saved in the working directory.}

\item{exclude}{A string or vector with the names of pages, which should be excluded from report assembly.}

\item{render}{A logical indicating whether the assembled report should immediately be rendered with \code{rmarkdown::render}.}

\item{...}{Additional arguments passed on to \code{rmarkdown::render}.}

\item{map}{A character vector containing colors and possible the levels they map to (as names).}

\item{name}{A name for the color mapping.}

\item{page}{The name of the page to be added.}

\item{title}{The title of the page to be added.}

\item{layout}{The page layout (see below).}

\item{menu}{The name of the menu, under which the page should appear.}

\item{content}{The name of a function of a path to a file.}

\item{global}{Whether or not to add the content to the global sidebar.}

\item{copy}{Whether or not to copy images to \code{dashboard@datadir}.}

\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{page}{The name of the page to be removed.}

\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{page}{The name of the page to which add the sidebar.}

\item{...}{Additional parameters passed to the components render function.}
}
\description{
This method allows to add content either to the global sidebar, or to a sidebar of an existing page.
}
