% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R, R/colormap.R, R/get_set.R,
%   R/pages.R, R/sidebar.R
\name{assemble,i2dashboard-method}
\alias{assemble,i2dashboard-method}
\alias{add_colormap,i2dashboard-method}
\alias{add_link,i2dashboard-method}
\alias{add_page,i2dashboard-method}
\alias{remove_page,i2dashboard-method}
\alias{add_to_sidebar,i2dashboard-method}
\title{Method to assemble a dashboard to a Rmd file.}
\usage{
\S4method{assemble}{i2dashboard}(
  dashboard,
  pages = names(dashboard@pages),
  file = dashboard@file,
  exclude = NULL,
  render = FALSE,
  ...
)

\S4method{add_colormap}{i2dashboard}(dashboard, map, name)

\S4method{add_link}{i2dashboard}(
  dashboard,
  href,
  title = NULL,
  icon = NULL,
  align = c("right", "left"),
  target = NULL
)

\S4method{add_page}{i2dashboard}(
  dashboard,
  page,
  title,
  layout = "default",
  menu = NULL,
  sidebar = NULL,
  ...
)

\S4method{remove_page}{i2dashboard}(dashboard, page)

\S4method{add_to_sidebar}{i2dashboard}(
  dashboard,
  content,
  page = "default",
  global = FALSE,
  copy = FALSE,
  ...
)
}
\arguments{
\item{dashboard}{A \linkS4class{i2dash::i2dashboard}.}

\item{pages}{A string or vector with the names of pages, which should be assembled to a dashboard.}

\item{file}{The output filename (recommend that the suffix should be '.Rmd'). This file will be saved in the working directory.}

\item{exclude}{A string or vector with the names of pages, which should be excluded from dashboard assembly.}

\item{render}{A logical indicating whether the assembled dashboard should immediately be rendered with \code{rmarkdown::render} or run with \code{rmarkdown::run}.}

\item{...}{Additional parameters passed to the components render function.}

\item{map}{A character vector containing colors and possible the levels they map to (as names).}

\item{name}{A name for the color mapping.}

\item{href}{The target of the link.}

\item{title}{The title of the page to be added.}

\item{icon}{An optional link icon (see https://rmarkdown.rstudio.com/flexdashboard/using.html#icon-sets)}

\item{align}{Optional argument that can be “left” or “right” (defaults = “right”) defining the alignment of the links in the navigation bar}

\item{target}{An optional target (e.g. "_blank")}

\item{page}{The name of the page to which add the sidebar.}

\item{layout}{The page layout (see below).}

\item{menu}{The name of the menu, under which the page should appear.}

\item{content}{The name of a function of a path to a file.}

\item{global}{Whether or not to add the content to the global sidebar.}

\item{copy}{Whether or not to copy images to \code{dashboard@datadir}.}
}
\description{
This method allows to add content either to the global sidebar, or to a sidebar of an existing page.
}
