% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2dashboard.R
\docType{class}
\name{idashboard-class}
\alias{idashboard-class}
\alias{i2dashboard}
\title{The idashboard S4 class.}
\usage{
\S4method{initialize}{i2dashboard}(.Object, ...)

\S4method{show}{i2dashboard}(object)

i2dashboard(...)
}
\value{
An i2dashboard object.
}
\description{
The idashboard S4 class.

Constructor method of the i2dashboard class.

Show method of the i2dashboard class.

Create a new i2dashboard object.
}
\section{Slots}{

\describe{
\item{\code{title}}{The dashboards title}

\item{\code{author}}{The author of the dashboard}

\item{\code{interactive}}{If a shiny-based report should be created}

\item{\code{theme}}{The theme of the dashboard}

\item{\code{datadir}}{Path to the directory, where report data is stored.}

\item{\code{file}}{The output filename (recommend that the suffix should be '.Rmd').}

\item{\code{pages}}{A list of dashboard pages}

\item{\code{sidebar}}{Content of the global sidebar}

\item{\code{colormaps}}{A named list with color mappings.}

\item{\code{source}}{Either a logical value describing whether the source code should be embeded through an item in the navigation bar or a link to a URL where the source code can be found online.}

\item{\code{social}}{A vector with any number of the following services: “facebook”, “twitter”, “google-plus”, “linkedin”, and “pinterest”. You can also specify “menu” to provide a generic sharing drop-down menu that includes all of the services.}

\item{\code{navbar}}{A list of links in the navigation bar (see https://rmarkdown.rstudio.com/flexdashboard/using.html#navigation_bar).}
}}

