% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2dashboard.R
\docType{class}
\name{idashboard-class}
\alias{idashboard-class}
\alias{i2dashboard}
\title{The idashboard S4 class.}
\usage{
\S4method{initialize}{i2dashboard}(.Object, ...)

\S4method{show}{i2dashboard}(object)

i2dashboard(...)
}
\description{
The idashboard S4 class.

Constructor method of the i2dashboard class.

Show method of the i2dashboard class.

Create a new i2dashboard object.
}
\section{Slots}{

\describe{
\item{\code{title}}{The dashboards title}

\item{\code{author}}{The author of the dashboard}

\item{\code{interactive}}{If a shiny-based report should be created}

\item{\code{theme}}{The theme of the dashboard}

\item{\code{datadir}}{Path to the directory, where report data is stored.}

\item{\code{file}}{Filename of the resulting report Rmd file.}

\item{\code{pages}}{A list of dashboard pages}
}}

