% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R, R/components.R, R/i2dashboard.R,
%   R/pages.R
\docType{methods}
\name{assemble,i2dashboard-method}
\alias{assemble,i2dashboard-method}
\alias{add_component,i2dashboard-method}
\alias{idashboard-class}
\alias{add_page,i2dashboard-method}
\alias{remove_page,i2dashboard-method}
\title{Method to assemble a dashboard and write it to a file}
\usage{
\S4method{assemble}{i2dashboard}(object, output_file, pages, ...)

\S4method{add_component}{i2dashboard}(object, page = "default",
  component, ...)

\S4method{add_page}{i2dashboard}(object, page, title, layout = "default",
  menu = NULL, sidebar = NULL, ...)

\S4method{remove_page}{i2dashboard}(object, page)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{output_file}{The output filename (recommend that the suffix should be '.Rmd'). This file will be saved in the working directory.}

\item{pages}{A string or vector with the names of pages, which should be assemble to a report.}

\item{...}{Additional parameters passed to the components render function.}

\item{page}{The name of the page to add the component to.}

\item{component}{The name of the component.}

\item{title}{The title of the page to be added.}

\item{layout}{The page layout (see below).}

\item{menu}{The name of the menu, under which the page should appear.}

\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{...}{Additional parameters passed to the components render function.}

\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{page}{The name of the page to be added.}

\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{page}{The name of the page to be removed.}
}
\description{
Method to assemble a dashboard and write it to a file

Method to add a component to a page of an i2dashboard object

The idashboard S4 class

Method to add a page to an i2dashboard object

Method to remove a page to an i2dashboard object
}
\section{Slots}{

\describe{
\item{\code{title}}{The dashboards title}

\item{\code{author}}{The author of the dashboard}

\item{\code{interactive}}{If a shiny-based report should be created}

\item{\code{theme}}{The theme of the dashboard}

\item{\code{pages}}{A list of dashboard pages}
}}

