% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/i2dashboard.R, R/pages.R
\docType{methods}
\name{add_component,i2dashboard-method}
\alias{add_component,i2dashboard-method}
\alias{idashboard-class}
\alias{i2dashboard}
\alias{add_page,i2dashboard-method}
\title{Method to add a component to a page of an i2dashboard object}
\usage{
\S4method{add_component}{i2dashboard}(object, page, component, ...)

\S4method{add_page}{i2dashboard}(object, page, title,
  layout = "storyboard", menu = title, ...)
}
\arguments{
\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{page}{The name of the page to add the component to.}

\item{component}{The name of the component.}

\item{...}{Additional parameters passed to the components render function.}

\item{title}{The title of the page to be added.}

\item{layout}{The page layout (see below).}

\item{menu}{The name of the menu, under which the page should appear.}

\item{object}{A \linkS4class{i2dash::i2dashboard} object.}

\item{page}{The name of the page to be added.}
}
\description{
Method to add a component to a page of an i2dashboard object

The idashboard S4 class

Method to add a page to an i2dashboard object
}
\section{Slots}{

\describe{
\item{\code{title}}{The dashboards title}

\item{\code{pages}}{A list of dashboard pages}
}}

