% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{marker}
\alias{marker}
\title{marker module server logic}
\usage{
marker(input, output, session, clarion)
}
\arguments{
\item{input}{Shiny's input object.}

\item{output}{Shiny's output object.}

\item{session}{Shiny's session object.}

\item{clarion}{A clarion object. See \code{\link[wilson]{Clarion}}. (Supports reactive)}
}
\value{
A named list containing reactives (highlight, color, labelColumn, label, clarion).
}
\description{
marker module server logic
}
