% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{pca module server logic}
\usage{
pca(input, output, session, data, types, levels = NULL, entryLabel = NULL,
  width = 28, height = 28, ppi = 72, scale = 1)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{data}{data.table data visualized in plot. (Supports Reactive)}

\item{types}{data.table: (Supports reactive)
column1: colnames of data
column2: corresponding column typ
column3: label (optional, used instead of id)
column4: sub_label (optional, added to id/ label)}

\item{levels}{Levels from which data is selected (Defaults to unique(metadata[["level"]])). (Supports Reactive)}

\item{entryLabel}{Define additional columns added to each entry (Default = NULL). Use a vector containing the desired columnnames e.g. c("column1", "column2").}

\item{width}{Width of the plot in cm. Defaults to 28 and supports reactive.}

\item{height}{Height of the plot in cm. Defaults to 28 and supports reactive.}

\item{ppi}{Pixel per inch. Defaults to 72 and supports reactive.}

\item{scale}{Scale plot size. Defaults to 1, supports reactive.}
}
\value{
A reactive containing list with dimensions.
}
\description{
pca module server logic
}
\details{
Width/ height/ ppi less or equal to zero will use default value.
}
