% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{render_image}
\alias{render_image}
\title{Method to embed an image file in a component}
\usage{
render_image(image, image_alt_text = NULL, title = NULL)
}
\arguments{
\item{image}{The path to the image file.}

\item{image_alt_text}{The alt text of the image.}

\item{title}{The components title.}
}
\value{
A character string containing the evaluated component
}
\description{
Method to embed an image file in a component
}
